Add boilerplate homepage
The homepage for a website is the index.html file which is served by
default by the web server when we visit our website's domain name. So
now when we visit our website domain name we get the boilerplate code
added in the index.html file. If this index.html file is missing we
won't be able to see any content on our website's main url and instead
we would get some sort of an error since server's are configured to by
default serve the index.html for the main domain name url.