Create a recipe folder containing pav-bhaji page
Adding a recipe directory to our codebase inside which we will store
the webpage's for different kinds of recipe. Here we have added a basic
boilerplate code for the pav-bhaji's recipe page. The name for this
pav-bhaji's recipe page is not index.html as index.html means that it
is the homepage so for any other webpage's which are not the homepage
we add name to the files which best describe them. Here the webpage's
name is all lowercase and also we do not use spaces while naming the
file instead of using spaces we would use the hypen(-) or underscore(_)
for separating the texts.