Add link to pav bhaji recipe page in homepage
To able be able to move from one webpage to another we make use of the
anchor tags. Here we have added link to our pav bhaji's recipe page to
our home page using the anchor (<a>) tag so that we can move to the
pav-bhaji's recipe page from the homepage just by clicking the
pav bhaji link.