Add width and height for the recipe page's image
Without this width and height attributes mentioned on the img tag when
the image is loaded in the webpage it shifts the content on our webpage
since initially while loading the image it has no height and width and
only when the image is loaded it sets the width and height of the image
tag based on the loaded image. This causes the content which was
rendered near it to shift when the image is loaded to make the space
available for the image, this is because initially the image had no
width and height so it started rendering the next contents of the
webpage right next to it and finally when the image was loaded which
has a defined width and height so to make space available for it to be
rendered on the page it shifts the contents besides it. This causes our
webpage to have a poor UX experience because of the jumping of our
contents on the page. It also effects our SEO as search engines rank
pages which doesn't have this page layout shift higher in its list.

So by adding the width and height attributes to the img tag we tell the
browser to render the layout of the img tag on the screen with the
provided size so even when the image is being loaded in the background
the browser reserve that space for the image and when the image is
loaded it fills that space thus having a no page layout shift and
providing a better UX.